+++ lib/CHANGELOG.md
@@ -4,11 +4,18 @@
 
 # 6.0.0-beta.1 (COMING SOON)
 
-**BREAKING CHANGE: Update to Angular/RxJS/NgRX v6**
+### BREAKING CHANGES: Update to Angular/RxJS/NgRX v6
 
 * Converted to RxJS v6.0.1 _without the compat library_
 * Converted to Angular v6.0.0
 * Converted to NgRX v6.0.0-beta.1
+* Update no longer uses upsert (because `@ngrx/entity` upsert no longer takes an `Update<T>`).
+* Can no longer change the pk via upsert (same reason).
+
+### Features
+
+* Add `ENTITY_CACHE_METAREDUCER`
+* Add `INITIAL_ENTITY_CACHE_STATE`
 
 <hr>